move stream event handlers to http
